delete anggaran done, show summary data done

+++ app/views/user/index.php
@@ -23,23 +23,19 @@
             <div class="card">
                 <div class="card-body">
                     <h4 class="mt-0 header-title">Input Data User</h4>
-                    <form action="<?= BASEURL; ?>/user/tambah" method="post" class="form-enter" onsubmit=""
-                        id="formInsertData">
+                    <form action="<?= BASEURL; ?>/user/tambah" method="post" class="form-enter" onsubmit="" id="formInsertData">
                         <!-- SELECT `id`, `user_name`, `password`, `user_type`, `no_pegawai` FROM `user` WHERE 1 -->
                         <div class="form-group row">
                             <label for="example-text-input" class="col-sm-2 col-form-label">User Name</label>
                             <div class="col-sm-10">
-                                <input class="form-control" type="hidden" value="" id="id_user" name="id"
-                                    placeholder="user name">
-                                <input class="form-control" type="text" value="" id="user_name" name="user_name"
-                                    placeholder="user name">
+                                <input class="form-control" type="hidden" value="" id="id_user" name="id" placeholder="user name">
+                                <input class="form-control" type="text" value="" id="user_name" name="user_name" placeholder="user name">
                             </div>
                         </div>
                         <div class="form-group row">
                             <label for="example-text-input" class="col-sm-2 col-form-label">Password</label>
                             <div class="col-sm-10">
-                                <input class="form-control" type="text" value="" id="password" name="password"
-                                    placeholder="password">
+                                <input class="form-control" type="text" value="" id="password" name="password" placeholder="password">
                             </div>
                         </div>
                         <div class="form-group row">
@@ -49,7 +45,7 @@
                                 <select class="form-control" id="user_type" name="user_type">
                                     <option value="">Select Type</option>
                                     <option value="<?= KEPALA_USR ?>">PPTK</option>
-                                    <option value="<?= BENDAHAR_USR ?>">Bendahara</option>
+                                    <option value="<?= BENDAHARA_USR ?>">Bendahara</option>
                                     <option value="<?= ADMIN_USR ?>">Admin</option>
                                 </select>
                             </div>
@@ -57,14 +53,11 @@
                         <div class="form-group row">
                             <label for="example-text-input" class="col-sm-2 col-form-label">No Pegawai</label>
                             <div class="col-sm-8">
-                                <input class="form-control" type="text" value="" id="no_pegawai"
-                                    placeholder="nomor pegawai" readonly>
-                                <input class="form-control" type="hidden" value="" id="no_pegawai_hide"
-                                    name="no_pegawai" placeholder="nomor pegawai">
+                                <input class="form-control" type="text" value="" id="no_pegawai" placeholder="nomor pegawai" readonly>
+                                <input class="form-control" type="hidden" value="" id="no_pegawai_hide" name="no_pegawai" placeholder="nomor pegawai">
                             </div>
                             <div class="col-sm-2">
-                                <button class="btn btn-primary waves-effect waves-light" type="button"
-                                    data-toggle="modal" data-target="#dataModal"> search </button>
+                                <button class="btn btn-primary waves-effect waves-light" type="button" data-toggle="modal" data-target="#dataModal"> search </button>
                             </div>
                         </div>
                         <div class="form-group row">
@@ -72,8 +65,7 @@
                             <div class="col-sm-5">
                                 <a href="#" class="btn btn-primary waves-effect waves-light" onclick="save_data();">
                                     Save </a>
-                                <button class="btn btn-danger waves-effect waves-light" type="reset"
-                                    onclick="reload_location('user')"> Reset </button>
+                                <button class="btn btn-danger waves-effect waves-light" type="reset" onclick="reload_location('user')"> Reset </button>
                             </div>
                         </div>
                     </form>
@@ -99,31 +91,28 @@
                             $dataPegawai    = $data['pegawai'];
                             $dataUser       = $data['user'];
                             foreach ($dataUser as $data) : ?>
-                            <tr>
-                                <td><?= $data['user_name']; ?></td>
-                                <td><?= $data['user_type']; ?></td>
-                                <td><?= $data['nama_pegawai']; ?></td>
-                                <td>
-                                    <a href="<?= BASEURL; ?>/user/hapus/<?= $data['id']; ?>"
-                                        class="btn btn-danger waves-effect waves-light"
-                                        onclick="return confirm('Yakin?');">
-                                        <span>
-                                            Hapus
-                                        </span>
-                                    </a>
-                                    <a href="#" class="getUbah btn btn-primary waves-effect waves-light"
-                                        data-id="<?= $data['id']; ?>">
-                                        <span>
-                                            Ubah
-                                        </span>
-                                    </a>
-                                    <!-- <a href="<?= BASEURL; ?>/user/detail/<?= $data['id']; ?>" class="">
+                                <tr>
+                                    <td><?= $data['user_name']; ?></td>
+                                    <td><?= $data['user_type']; ?></td>
+                                    <td><?= $data['nama_pegawai']; ?></td>
+                                    <td>
+                                        <a href="<?= BASEURL; ?>/user/hapus/<?= $data['id']; ?>" class="btn btn-danger waves-effect waves-light" onclick="return confirm('Yakin?');">
+                                            <span>
+                                                Hapus
+                                            </span>
+                                        </a>
+                                        <a href="#" class="getUbah btn btn-primary waves-effect waves-light" data-id="<?= $data['id']; ?>">
+                                            <span>
+                                                Ubah
+                                            </span>
+                                        </a>
+                                        <!-- <a href="<?= BASEURL; ?>/user/detail/<?= $data['id']; ?>" class="">
                                         <span>
                                             Detail
                                         </span>
                                     </a>  -->
-                                </td>
-                            </tr>
+                                    </td>
+                                </tr>
                             <?php endforeach; ?>
                         </tbody>
                     </table>
@@ -135,8 +124,7 @@
 </div>
 
 <!-- Modal -->
-<div class="modal fade bd-example-modal-lg" id="dataModal" tabindex="-1" role="dialog" aria-labelledby="dataModalLabel"
-    aria-hidden="true">
+<div class="modal fade bd-example-modal-lg" id="dataModal" tabindex="-1" role="dialog" aria-labelledby="dataModalLabel" aria-hidden="true">
     <div class="modal-dialog modal-lg" role="document">
         <div class="modal-content">
             <div class="modal-header">
@@ -158,19 +146,18 @@
                     <tbody>
                         <?php $no = 1; ?>
                         <?php foreach ($dataPegawai as $data) : ?>
-                        <tr>
-                            <td><?= $no; ?></td>
-                            <td>
-                                <a href="#" class="getNomorPegawai" data-nomor="<?= $data['no_pegawai']; ?>"
-                                    data-dismiss="modal">
-                                    <span>
-                                        <?= $data['no_pegawai']; ?>
-                                    </span>
-                                </a>
-                            </td>
-                            <td><?= $data['nama_pegawai']; ?></td>
-                        </tr>
-                        <?php $no++; ?>
+                            <tr>
+                                <td><?= $no; ?></td>
+                                <td>
+                                    <a href="#" class="getNomorPegawai" data-nomor="<?= $data['no_pegawai']; ?>" data-dismiss="modal">
+                                        <span>
+                                            <?= $data['no_pegawai']; ?>
+                                        </span>
+                                    </a>
+                                </td>
+                                <td><?= $data['nama_pegawai']; ?></td>
+                            </tr>
+                            <?php $no++; ?>
                         <?php endforeach; ?>
                     </tbody>
                 </table>
@@ -179,87 +166,95 @@
     </div>
 
     <script>
-    $(document).ready(function() {
-        $('.form-enter').on('keypress', function(e) {
-            return e.which !== 13;
-        });
+        $(document).ready(function() {
+            $('.form-enter').on('keypress', function(e) {
+                return e.which !== 13;
+            });
 
-        $('.data-table-format').DataTable();
+            $('.data-table-format').DataTable();
 
-        $('.getUbah').on('click', function() {
-            const id = $(this).data('id')
-            console.log(id);
-            $.ajax({
-                url: '<?= BASEURL; ?>/user/getUbah/',
-                data: {
-                    id: id
-                },
-                method: 'post',
-                dataType: 'json',
-                success: function(data) {
-                    console.log(data);
-                    $('#id_user').val(data.id);
-                    $('#user_name').val(data.user_name);
-                    $('#password').val(data.password);
-                    $('#user_type').val(data.user_type);
-                    $('#no_pegawai').val(data.no_pegawai);
-                    $('#no_pegawai_hide').val(data.no_pegawai);
+            $('.getUbah').on('click', function() {
+                const id = $(this).data('id')
+                console.log(id);
+                $.ajax({
+                    url: '<?= BASEURL; ?>/user/getUbah/',
+                    data: {
+                        id: id
+                    },
+                    method: 'post',
+                    dataType: 'json',
+                    beforeSend: function() {
+                        $.blockUI({
+                            message: null
+                        });
+                    },
+                    complete: function() {
+                        $.unblockUI();
+                    },
+                    success: function(data) {
+                        console.log(data);
+                        $('#id_user').val(data.id);
+                        $('#user_name').val(data.user_name);
+                        $('#password').val(data.password);
+                        $('#user_type').val(data.user_type);
+                        $('#no_pegawai').val(data.no_pegawai);
+                        $('#no_pegawai_hide').val(data.no_pegawai);
 
-                    $(".card-body form").attr('action', '<?= BASEURL; ?>/user/ubah')
-                    $('.card-body form button[type=submit]').html('Ubah Data')
-                },
-                error: function(){
-                    console.log("GAGAL");
-                }
+                        $(".card-body form").attr('action', '<?= BASEURL; ?>/user/ubah')
+                        $('.card-body form button[type=submit]').html('Ubah Data')
+                    },
+                    error: function() {
+                        console.log("GAGAL");
+                    }
+                })
+            })
+
+            $('.getNomorPegawai').on('click', function() {
+                const nomor = $(this).data('nomor');
+                $('#no_pegawai').val(nomor)
+                $('#no_pegawai_hide').val(nomor)
+                $('#dataModal').modal('no_pegawai');
+                $('#dataModal').modal('no_pegawai_hide');
             })
         })
 
-        $('.getNomorPegawai').on('click', function() {
-            const nomor = $(this).data('nomor');
-            $('#no_pegawai').val(nomor)
-            $('#no_pegawai_hide').val(nomor)
-            $('#dataModal').modal('no_pegawai');
-            $('#dataModal').modal('no_pegawai_hide');
-        })
-    })
+        function save_data() {
+            if ($('#user_name').val() == "") {
+                $("#message").html(message('gagal', 'diubah atau ditambahkan, data yang di isi harus lengkap', 'danger',
+                    'User'));
+                return
 
-    function save_data() {
-        if ($('#user_name').val() == "") {
-            $("#message").html(message('gagal', 'diubah atau ditambahkan, data yang di isi harus lengkap', 'danger',
-                'User'));
-            return
+            }
+            if ($('#password').val() == "") {
+                $("#message").html(message('gagal', 'diubah atau ditambahkan, data yang di isi harus lengkap', 'danger',
+                    'User'));
 
-        }
-        if ($('#password').val() == "") {
-            $("#message").html(message('gagal', 'diubah atau ditambahkan, data yang di isi harus lengkap', 'danger',
-                'User'));
+                return
 
-            return
+            }
+            if ($('#user_type').val() == "") {
+                $("#message").html(message('gagal', 'diubah atau ditambahkan, data yang di isi harus lengkap', 'danger',
+                    'User'));
+                return
 
-        }
-        if ($('#user_type').val() == "") {
-            $("#message").html(message('gagal', 'diubah atau ditambahkan, data yang di isi harus lengkap', 'danger',
-                'User'));
-            return
+            }
+            if ($('#no_pegawai_hide').val() == "") {
+                $("#message").html(message('gagal', 'diubah atau ditambahkan, data yang di isi harus lengkap', 'danger',
+                    'User'));
+                return
+            }
 
-        }
-        if ($('#no_pegawai_hide').val() == "") {
-            $("#message").html(message('gagal', 'diubah atau ditambahkan, data yang di isi harus lengkap', 'danger',
-                'User'));
-            return
+            $('#formInsertData').submit();
         }
 
-        $('#formInsertData').submit();
-    }
-
-    function message(pesan, aksi, tipe, data) {
-        allert_load = "";
-        allert_load += '<div class="alert alert-' + tipe + ' alert-dismissible fade show" role="alert">'
-        allert_load += 'Data ' + data + ' <strong>' + pesan + ' </strong> ' + aksi
-        allert_load += '<button type="button" class="close" data-dismiss="alert" aria-label="Close">'
-        allert_load += '<span aria-hidden="true">&times;</span>'
-        allert_load += '</button>'
-        allert_load += '</div>'
-        return allert_load
-    }
+        function message(pesan, aksi, tipe, data) {
+            allert_load = "";
+            allert_load += '<div class="alert alert-' + tipe + ' alert-dismissible fade show" role="alert">'
+            allert_load += 'Data ' + data + ' <strong>' + pesan + ' </strong> ' + aksi
+            allert_load += '<button type="button" class="close" data-dismiss="alert" aria-label="Close">'
+            allert_load += '<span aria-hidden="true">&times;</span>'
+            allert_load += '</button>'
+            allert_load += '</div>'
+            return allert_load
+        }
     </script>
